+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - MySQL: Do not export IN in FUNCTION parameters (regression from 5.4.2)
 - MySQL: Convert vector columns to text in select and edit
 - MySQL, SQLite: Draggable columns in alter table
+- MariaDB 11.7+: Support vector columns and indexes
 - PostgreSQL 9.2+: Highlight queries in process list
 - MS SQL: Fix multiple result sets (bug #1180)
 - New plugin: Redis driver